Update update image of user to flutter app v1

+++ app/Services/Storage/StorageService.php
@@ -23,19 +23,17 @@ class StorageService
 
     public function replace(UploadedFile $newFile, ?string $oldPath,string $folder,string $disk = self::DISK_PUBLIC,?string $name = null)
     {
-        logger()->info('Avatar Replace', [
-            'old_path' => $oldPath,
-            'exists' => Storage::disk($disk)->exists($oldPath),
-        ]);
         $data = $this->upload($newFile, $folder, $disk, $name);
-        $this->delete($oldPath, $disk);
+        if (!empty($oldPath)) {
+            $this->delete($oldPath, $disk);
+        }
 
         return $data ;
     }
 
     public function delete(?string $path, string $disk = self::DISK_PUBLIC)
     {
-        if (! $path) {
+        if (!$path) {
             return false;
         }
         return Storage::disk($disk)->delete($path);
